test: continue splitting coordinator tests by behavior
