+++ tests/test_coordinator.py
@@ -212,59 +212,6 @@ async def test_read_input_registers_reconnect_on_error(coordinator):
 
 
 @pytest.mark.asyncio
-async def test_disconnect_retry_transportless_restores_client(coordinator):
-    """Transport-less retry restores client when disconnect clears it."""
-    client = MagicMock()
-    coordinator.client = client
-    coordinator._transport = None
-    coordinator._disconnect = _disconnect_clear_client(coordinator)
-    coordinator._ensure_connection = AsyncMock()
-
-    reconnect_error = await coordinator._disconnect_and_reconnect_for_retry(
-        register_type="input",
-        start_address=0,
-        attempt=1,
-    )
-
-    assert reconnect_error is None
-    assert coordinator.client is client
-    coordinator._ensure_connection.assert_not_awaited()
-
-
-@pytest.mark.asyncio
-async def test_disconnect_retry_transportless_returns_disconnect_error(coordinator):
-    """Transport-less retry returns disconnect error and skips reconnect."""
-    client = MagicMock()
-    coordinator.client = client
-    coordinator._transport = None
-    coordinator._disconnect = _disconnect_raise_oserror()
-    coordinator._ensure_connection = AsyncMock()
-
-    reconnect_error = await coordinator._disconnect_and_reconnect_for_retry(
-        register_type="input",
-        start_address=0,
-        attempt=1,
-    )
-
-    assert isinstance(reconnect_error, OSError)
-    coordinator._ensure_connection.assert_not_awaited()
-
-
-def _disconnect_clear_client(coordinator):
-    async def _disconnect() -> None:
-        coordinator.client = None
-
-    return _disconnect
-
-
-def _disconnect_raise_oserror():
-    async def _disconnect() -> None:
-        raise OSError("disconnect failed")
-
-    return _disconnect
-
-
-@pytest.mark.asyncio
 async def test_async_write_multi_register_start(coordinator, monkeypatch):
     """Writing multi-register from start address succeeds."""
     coordinator.async_request_refresh = AsyncMock()
